Fix unread counter
after_entry_id query parameter was not being used correctly. We have to get all
the entries every time in order to know the total number of unread. "total" in
API repsone is total included in response, not the total unread altogether.

Also, fix wording for single notifications of multiple new items